requirements.txt
appnope==0.1.0
arrow==0.6.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.3.0
Django==1.9.2
django-allauth==0.24.1
django-environ==0.4.0
django-extensions==1.6.7
django-filter==0.13.0
django-rest-auth==0.6.0
django-rest-swagger==0.3.7
djangorestframework==3.3.2
environ==1.0
gnureadline==6.3.3
gunicorn==19.4.5
ipython==4.2.0
ipython-genutils==0.1.0
oauthlib==1.0.3
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
ptyprocess==0.5.1
python-dateutil==2.5.3
python-openid==2.2.5
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
tapioca-facebook==0.6.1
tapioca-wrapper==1.2.1
traitlets==4.2.1
Unidecode==0.4.19
whitenoise==2.0.6
